+++ tests/unit/chrono/TimeTest.php
@@ -53,6 +53,23 @@ class TimeTest extends TestCase
 	/**
 	 *
 	 */
+	public function testToNative(): void
+	{
+		$time1 = new Time('now', 'Europe/Oslo');
+
+		$time2 = $time1->toNative();
+
+		$this->assertInstanceOf(DateTime::class, $time2);
+		$this->assertNotInstanceOf(Time::class, $time2);
+
+		$this->assertSame($time1->getTimestamp(), $time2->getTimestamp());
+
+		$this->assertSame($time1->getTimezone()->getName(), $time2->getTimezone()->getName());
+	}
+
+	/**
+	 *
+	 */
 	public function testConstructor(): void
 	{
 		$time = new Time;