+++ CHANGELOG.md
@@ -27,6 +27,8 @@ The major version bump is due to upping the required PHP version from `8.5` to `
 * The development error handler now allows you to inspect the stack traces of previous exceptions.
 * The development error handler may now provide hints about deprecations.
 * Added `TimeInterface::createFromFormatOrThrow()` method.
+* Added `Time::toNative()` method.
+* Added `TimeImmutable::toNative()` method.
 * Added `PSR-20` (ClockInterface) implementation.
 * Added `SleeperInterface` and `Sleeper` implementation for time-based delays.